+++ build-logic/src/main/kotlin/HtmlValidator.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2024 Apple Inc. and the Pkl project authors. All rights reserved.
+ * Copyright © 2024-2026 Apple Inc. and the Pkl project authors. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import org.gradle.util.GradleVersion
+import org.gradle.api.Project
+import org.gradle.api.file.FileCollection
 
-open class GradlePluginTests {
-  lateinit var minGradleVersion: GradleVersion
-  lateinit var maxGradleVersion: GradleVersion
-  var skippedGradleVersions: List<GradleVersion> = listOf()
+open class HtmlValidator(project: Project) {
+  var sources: FileCollection = project.files()
 }
