+++ build-logic/src/main/kotlin/pklJavaLibrary.gradle.kts
@@ -56,7 +56,9 @@ spotless {
     addStep(revertYearOnlyChanges)
     googleJavaFormat(libs.versions.googleJavaFormat.get())
     target("src/*/java/**/*.java")
-    licenseHeaderFile(rootProject.file("buildSrc/src/main/resources/license-header.star-block.txt"))
+    licenseHeaderFile(
+      rootProject.file("build-logic/src/main/resources/license-header.star-block.txt")
+    )
   }
 }
