+++ pkl-commons-test/pkl-commons-test.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2024-2025 Apple Inc. and the Pkl project authors. All rights reserved.
+ * Copyright © 2024-2026 Apple Inc. and the Pkl project authors. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,8 @@
 import java.security.MessageDigest
 
 plugins {
-  pklAllProjects
-  pklKotlinLibrary
+  id("pklAllProjects")
+  id("pklKotlinLibrary")
 }
 
 // note: no need to publish this library
