+++ pkl-doc/pkl-doc.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2024-2025 Apple Inc. and the Pkl project authors. All rights reserved.
+ * Copyright © 2024-2026 Apple Inc. and the Pkl project authors. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 plugins {
-  pklAllProjects
-  pklKotlinLibrary
-  pklJavaExecutable
-  pklNativeExecutable
-  pklHtmlValidator
+  id("pklAllProjects")
+  id("pklKotlinLibrary")
+  id("pklJavaExecutable")
+  id("pklNativeExecutable")
+  id("pklHtmlValidator")
   alias(libs.plugins.kotlinxSerialization)
 }
