+++ stdlib/stdlib.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2024-2025 Apple Inc. and the Pkl project authors. All rights reserved.
+ * Copyright © 2024-2026 Apple Inc. and the Pkl project authors. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 plugins {
-  pklAllProjects
+  id("pklAllProjects")
   base
   `maven-publish`
-  pklPublishLibrary
-  pklSpotlessFormat
+  id("pklPublishLibrary")
+  id("pklSpotlessFormat")
   signing
 }
